Add files js and script on index.html
