+++ boissons.js
@@ -1,0 +1,3 @@
+const boissonsMenu = [ "Coca","Fanta","Sprite","Eau"]
+
+export default boissonsMenu;
