+++ desserts.js
@@ -1,0 +1,3 @@
+const dessertsMenu = [ "Tiramisu","Glace à la vanille","Glace au chocolat","Glace au café"]
+
+export default dessertsMenu;
