+++ entrees.js
@@ -1,0 +1,3 @@
+const entreesMenu = [ "Salade de tomates","Mozarella - Roquette","Chèvres chauds","Antipasti italiens"];
+
+export default entreesMenu;
