+++ index.html
@@ -26,7 +26,7 @@
 
     <div class="menu-jour">
         <h2>MENU DU JOUR</h2>
-        <h2>DATE</h2>
+        <h2><span id="date">DATE</span></h2>
         <p><span id="entrees">Entrées</span>
             <br/>***
         <br/><span id="plats">Plats</span>
@@ -46,6 +46,7 @@
         </div>
         <div class="icon-valeurs"></div>
     </div>
-    
+     
 </body>
+<script type="module" src="./index.js"></script>
 </html>
