+++ plats.js
@@ -1,0 +1,3 @@
+const platsMenu = [ "Pizza Buffala","Pâtes aux légumes du soleil","Escalope Milanaise","Canard à l'italienne"]
+
+export default platsMenu;
